requirements.txt
numpy==1.21
pandas==1.4
scipy==1.8
adjustText==0.7.3
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
filelock==3.4.1
gdown==4.7
idna==3.3
importlib-resources==5.4.0
kiwisolver==1.3.1
matplotlib==3.5
Pillow==8.4.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
seaborn==0.11.2
six==1.16.0
soupsieve==2.3.1
tqdm==4.63.1
urllib3==1.26.9
XlsxWriter==3.0.3
zipp==3.6.0
